Add files via upload
Basic structure of diaries and adding of new ones. Not much code in here yet but the structure seems to work. 
It's probably gonna be useful if we want to start working on stuff.
I haven't done javadoc comments yet but it should be pretty self-explanatory.
To add a diary, use sth like diarytree.add(new Diary("Firstname", "Lastname");
A simple way of using a diary is not implemented properly yet. The tree contains linkedlists of diaries so you can use those.
Searching by name as of right now is done by creating a new list of one diary with this name and using diarytree.search(theList);